* develop:
  Add callback macro workaround for gcc 4.3.3
  Add empty virtualMidi test to test suite
  Add VirtualMidi class implementation
  Fix warnings for C++11
  Add c++0x flag for test suite
  Update flags so that g++-4.7 works
  Define callback_ptr_t directly in MidiCallback class
  Add callback helper macros using decltype

+++ hardware/libraries/MNM/MNM.cpp
@@ -134,7 +134,7 @@ void MNMClass::sendSysex(uint8_t *bytes, uint8_t cnt) {
 }
 
 void MNMClass::setStatus(uint8_t id, uint8_t value) {
-  uint8_t data[] = { 0x71, id & 0x7F, value & 0x7F };
+  uint8_t data[] = { 0x71, (uint8_t)(id & 0x7F), (uint8_t)(value & 0x7F) };
   MNM.sendSysex(data, countof(data));
 }
